# internal/docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "VATUSAs internal API for use by other internal VATUSA services.",
        "title": "VATUSA Internal",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "VATUSA Support",
            "url": "http://www.swagger.io/support",
            "email": "vatusa6@vatusa.net"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "0.1"
    },
    "host": "api.vatusa.local",
    "basePath": "/internal/v1",
    "paths": {
...
        "/documents/facility/{Facility}": {
            "get": {
                "description": "List all documents by facility",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "documents"
                ],
                "summary": "List all documents by facility",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Facility ID",
                        "name": "Facility",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/document.Response"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrResponse"
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrResponse"
                        }
                    }
                }
            }
        },
...
        "/facility-log": {
            "get": {
                "description": "List facility log entries",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "facility-log"
                ],
                "summary": "List facility log entries",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/facility_log.Response"
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "Create a new facility log entry",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "facility-log"
                ],
                "summary": "Create a new facility log entry",
                "parameters": [
                    {
                        "description": "Facility Log Entry",
                        "name": "facility_log",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/facility_log.Request"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/facility_log.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrResponse"
                        }
                    }
                }
            }
        },
        "/facility-log/{id}": {
            "get": {
                "description": "Get a facility log entry",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "facility-log"
                ],
                "summary": "Get a facility log entry",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Facility Log Entry ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/facility_log.Response"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrResponse"
                        }
                    }
                }
            },
            "put": {
                "description": "Update a facility log entry",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "facility-log"
                ],
                "summary": "Update a facility log entry",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Facility Log Entry ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Facility Log Entry",
                        "name": "facility_log",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/facility_log.Request"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/facility_log.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrResponse"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete a facility log entry",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "facility-log"
                ],
                "summary": "Delete a facility log entry",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Facility Log Entry ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrResponse"
                        }
                    }
                }
            },
            "patch": {
                "description": "Patch a facility log entry",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "facility-log"
                ],
                "summary": "Patch a facility log entry",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Facility Log Entry ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Facility Log Entry",
                        "name": "facility_log",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/facility_log.Request"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/facility_log.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrResponse"
                        }
                    }
                }
            }
        },
...
        "/feedback": {
            "get": {
                "description": "List feedback entries",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "feedback"
                ],
                "summary": "List feedback entries",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/feedback.Response"
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "Create a new feedback entry",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "feedback"
                ],
                "summary": "Create a new feedback entry",
                "parameters": [
                    {
                        "description": "Feedback Entry",
                        "name": "feedback",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/feedback.Request"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/feedback.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrResponse"
                        }
                    }
                }
            }
        },
        "/feedback/{id}": {
            "get": {
                "description": "Get a feedback entry",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "feedback"
                ],
                "summary": "Get a feedback entry",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Feedback ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/feedback.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrResponse"
                        }
                    }
                }
            },
            "put": {
                "description": "Update a feedback entry",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "feedback"
                ],
                "summary": "Update a feedback entry",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Feedback ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Feedback Entry",
                        "name": "feedback",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/feedback.Request"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrResponse"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete a feedback entry",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "feedback"
                ],
                "summary": "Delete a feedback entry",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Feedback ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrResponse"
                        }
                    }
                }
            },
            "patch": {
                "description": "Patch a feedback entry",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "feedback"
                ],
                "summary": "Patch a feedback entry",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Feedback ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Feedback Entry",
                        "name": "feedback",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/feedback.Request"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "document.Response": {
            "type": "object",
            "properties": {
                "category": {
                    "allOf": [
                        {
                            "$ref": "#/definitions/types.DocumentCategory"
                        }
                    ],
                    "example": "general"
                },
                "created_at": {
                    "type": "string",
                    "example": "2021-01-01T00:00:00Z"
                },
                "created_by": {
                    "type": "integer",
                    "example": 1293257
                },
                "description": {
                    "type": "string",
                    "example": "General Division Policy"
                },
                "facility": {
                    "type": "string",
                    "example": "ZDV"
                },
                "id": {
                    "type": "integer",
                    "example": 1
                },
                "name": {
                    "type": "string",
                    "example": "DP001"
                },
                "updated_at": {
                    "type": "string",
                    "example": "2021-01-01T00:00:00Z"
                },
                "updated_by": {
                    "type": "integer",
                    "example": 1293257
                },
                "url": {
                    "type": "string",
                    "example": "https://zdvartcc.org"
                }
            }
        },
        "facility_log.Request": {
            "type": "object",
            "required": [
                "entry",
                "facility"
            ],
            "properties": {
                "entry": {
                    "type": "string",
                    "example": "Changed Preferred OIs to RP"
                },
                "facility": {
                    "type": "string",
                    "example": "ZDV"
                }
            }
        },
        "facility_log.Response": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string",
                    "example": "2021-01-01T00:00:00Z"
                },
                "created_by": {
                    "type": "string",
                    "example": "'1234567' or 'System'"
                },
                "entry": {
                    "type": "string",
                    "example": "Change URL from 'denartcc.org' to 'zdvartcc.org'"
                },
                "facility": {
                    "type": "string",
                    "example": "ZDV"
                },
                "id": {
                    "type": "integer",
                    "example": 1
                },
                "updated_at": {
                    "type": "string",
                    "example": "2021-01-01T00:00:00Z"
                },
                "updated_by": {
                    "type": "string",
                    "example": "'1234567' or 'System'"
                }
            }
        },
...
        "feedback.Request": {
            "type": "object",
            "required": [
                "callsign",
                "controller_cid",
                "facility",
                "notes",
                "pilot_cid",
                "position",
                "rating",
                "status"
            ],
            "properties": {
                "callsign": {
                    "type": "string",
                    "example": "DAL123"
                },
                "comment": {
                    "type": "string",
                    "example": "Great work Raaj!"
                },
                "controller_cid": {
                    "type": "integer",
                    "example": 1293257
                },
                "facility": {
                    "type": "string",
                    "example": "ZDV"
                },
                "notes": {
                    "type": "string",
                    "example": "Raaj was the best controller I've ever flown under."
                },
                "pilot_cid": {
                    "type": "integer",
                    "example": 1293257
                },
                "position": {
                    "type": "string",
                    "example": "DEN_I_APP"
                },
                "rating": {
                    "enum": [
                        "unsatisfactory",
                        "poor",
                        "fair",
                        "good",
                        "excellent"
                    ],
                    "allOf": [
                        {
                            "$ref": "#/definitions/types.FeedbackRating"
                        }
                    ],
                    "example": "good"
                },
                "status": {
                    "enum": [
                        "pending",
                        "approved",
                        "denied"
                    ],
                    "allOf": [
                        {
                            "$ref": "#/definitions/types.StatusType"
                        }
                    ],
                    "example": "pending"
                }
            }
        },
        "feedback.Response": {
            "type": "object",
            "properties": {
                "callsign": {
                    "type": "string",
                    "example": "DAL123"
                },
                "comment": {
                    "type": "string",
                    "example": "Great work Raaj!"
                },
                "controller": {
                    "$ref": "#/definitions/models.User"
                },
                "created_at": {
                    "type": "string",
                    "example": "2021-01-01T00:00:00Z"
                },
                "facility": {
                    "type": "string",
                    "example": "ZDV"
                },
                "id": {
                    "type": "integer",
                    "example": 1
                },
                "notes": {
                    "type": "string",
                    "example": "Raaj was the best controller I've ever flown under."
                },
                "pilot": {
                    "$ref": "#/definitions/models.User"
                },
                "position": {
                    "type": "string",
                    "example": "DEN_I_APP"
                },
                "rating": {
                    "allOf": [
                        {
                            "$ref": "#/definitions/types.FeedbackRating"
                        }
                    ],
                    "example": "1"
                },
                "status": {
                    "allOf": [
                        {
                            "$ref": "#/definitions/types.StatusType"
                        }
                    ],
                    "example": "pending"
                },
                "updated_at": {
                    "type": "string",
                    "example": "2021-01-01T00:00:00Z"
                }
            }
        },
...
        "models.User": {
            "type": "object",
            "properties": {
                "cid": {
                    "type": "integer",
                    "example": 1293257
                },
                "controller_rating": {
                    "type": "integer",
                    "example": 1
                },
                "created_at": {
                    "type": "string",
                    "example": "2021-01-01T00:00:00Z"
                },
                "discord_id": {
                    "type": "string",
                    "example": "1234567890"
                },
                "email": {
                    "type": "string",
                    "example": "vatusa6@vatusa.net"
                },
                "first_name": {
                    "type": "string",
                    "example": "Raaj"
                },
                "flags": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.UserFlag"
                    }
                },
                "last_cert_sync": {
                    "type": "string",
                    "example": "2021-01-01T00:00:00Z"
                },
                "last_login": {
                    "type": "string",
                    "example": "2021-01-01T00:00:00Z"
                },
                "last_name": {
                    "type": "string",
                    "example": "Patel"
                },
                "pilot_rating": {
                    "type": "integer",
                    "example": 1
                },
                "pref_name_enabled": {
                    "type": "boolean",
                    "example": true
                },
                "preferred_name": {
                    "type": "string",
                    "example": "Raaj"
                },
                "preferred_ois": {
                    "type": "string",
                    "example": "RP"
                },
                "roles": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.UserRole"
                    }
                },
                "updated_at": {
                    "type": "string",
                    "example": "2021-01-01T00:00:00Z"
                }
            }
        },
...
        "types.DocumentCategory": {
            "type": "string",
            "enum": [
                "general",
                "training",
                "information_technology",
                "sops",
                "loas",
                "misc"
            ],
            "x-enum-varnames": [
                "General",
                "Training",
                "InformationTechnology",
                "SOPs",
                "LOAs",
                "Misc"
            ]
        },
        "types.FeedbackRating": {
            "type": "string",
            "enum": [
                "unsatisfactory",
                "poor",
                "fair",
                "good",
                "excellent"
            ],
            "x-enum-varnames": [
                "Unsatisfactory",
                "Poor",
                "Fair",
                "Good",
                "Excellent"
            ]
        },
...
        "types.StatusType": {
            "type": "string",
            "enum": [
                "pending",
                "accepted",
                "rejected"
            ],
            "x-enum-varnames": [
                "Pending",
                "Accepted",
                "Rejected"
            ]
        },
...
        "utils.ErrResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "application-specific error code",
                    "type": "integer"
                },
                "error": {
                    "description": "application-level error message, for debugging",
                    "type": "string"
                },
                "status": {
                    "description": "user-level status message",
                    "type": "string"
                }
            }
        }
    }
}